Reorganizing features.cpp
